fix alt key system
I am still not quite satisfied with the alt key system.
Basically, it focus the docked menu when pressing the Alt Key (or the
key supplied with -a), so the docked menu can be controlled as a regular
menu.  It has no effect when the `-d` option is not used.